website/docs: Clean up config. Add types.
