--- website/sidebarsIntegrations.js
@@ -1,209 +1,0 @@
-module.exports = {
-    integrations: [
-        {
-            type: "doc",
-            id: "index",
-        },
-        {
-            type: "category",
-            label: "Applications",
-            collapsed: false,
-            link: {
-                type: "doc",
-                id: "services/index",
-            },
-            items: [
-                {
-                    type: "category",
-                    label: "Device Management",
-                    items: ["services/apple/index", "services/fleet/index"],
-                },
-                {
-                    type: "category",
-                    label: "Chat, Communication & Collaboration",
-                    items: [
-                        "services/bookstack/index",
-                        "services/dokuwiki/index",
-                        "services/espocrm/index",
-                        "services/hedgedoc/index",
-                        "services/kimai/index",
-                        "services/mailcow/index",
-                        "services/mastodon/index",
-                        "services/matrix-synapse/index",
-                        "services/mautic/index",
-                        "services/mobilizon/index",
-                        "services/nextcloud/index",
-                        "services/onlyoffice/index",
-                        "services/openproject/index",
-                        "services/outline/index",
-                        "services/owncloud/index",
-                        "services/paperless-ng/index",
-                        "services/paperless-ngx/index",
-                        "services/rocketchat/index",
-                        "services/roundcube/index",
-                        "services/sharepoint-se/index",
-                        "services/slack/index",
-                        "services/thelounge/index",
-                        "services/vikunja/index",
-                        "services/wekan/index",
-                        "services/wiki-js/index",
-                        "services/writefreely/index",
-                        "services/zulip/index",
-                    ],
-                },
-                {
-                    type: "category",
-                    label: "Cloud Providers",
-                    items: [
-                        "services/aws/index",
-                        "services/google/index",
-                        "services/hashicorp-cloud/index",
-                        "services/oracle-cloud/index",
-                    ],
-                },
-                {
-                    type: "category",
-                    label: "Dashboards",
-                    items: [
-                        "services/organizr/index",
-                        "services/linkwarden/index",
-                        "services/homarr/index",
-                    ],
-                },
-                {
-                    type: "category",
-                    label: "Hypervisors / Orchestrators",
-                    items: [
-                        "services/portainer/index",
-                        "services/proxmox-ve/index",
-                        "services/rancher/index",
-                        "services/xen-orchestra/index",
-                        "services/vmware-vcenter/index",
-                    ],
-                },
-                {
-                    type: "category",
-                    label: "Infrastructure",
-                    items: [
-                        "services/apache-guacamole/index",
-                        "services/argocd/index",
-                        "services/awx-tower/index",
-                        "services/cloudflare-access/index",
-                        "services/globalprotect/index",
-                        "services/harbor/index",
-                        "services/hashicorp-vault/index",
-                        "services/jenkins/index",
-                        "services/knocknoc/index",
-                        "services/meshcentral/index",
-                        "services/minio/index",
-                        "services/netbox/index",
-                        "services/omni/index",
-                        "services/pgadmin/index",
-                        "services/phpipam/index",
-                        "services/plesk/index",
-                        "services/powerdns-admin/index",
-                        "services/proftpd/index",
-                        "services/qnap-nas/index",
-                        "services/rustdesk-pro/index",
-                        "services/semgrep/index",
-                        "services/synology-dsm/index",
-                        "services/skyhigh/index",
-                        "services/snipe-it/index",
-                        "services/sssd/index",
-                        "services/terrakube/index",
-                        "services/truecommand/index",
-                        "services/veeam-enterprise-manager/index",
-                        "services/xcreds/index",
-                        "services/zammad/index",
-                    ],
-                },
-                {
-                    type: "category",
-                    label: "Networking",
-                    items: [
-                        "services/aruba-orchestrator/index",
-                        "services/firezone/index",
-                        "services/fortigate-admin/index",
-                        "services/fortigate-ssl/index",
-                        "services/fortimanager/index",
-                        "services/gravity/index",
-                        "services/netbird/index",
-                        "services/opnsense/index",
-                        "services/pfsense/index",
-                    ],
-                },
-                {
-                    type: "category",
-                    label: "Miscellaneous",
-                    items: [
-                        "services/actual-budget/index",
-                        "services/adventurelog/index",
-                        "services/calibre-web/index",
-                        "services/engomo/index",
-                        "services/frappe/index",
-                        "services/freshrss/index",
-                        "services/gravitee/index",
-                        "services/karakeep/index",
-                        "services/home-assistant/index",
-                        "services/immich/index",
-                        "services/jellyfin/index",
-                        "services/komga/index",
-                        "services/mealie/index",
-                        "services/miniflux/index",
-                        "services/node-red/index",
-                        "services/open-webui/index",
-                        "services/semaphore/index",
-                        "services/sonar-qube/index",
-                        "services/sonarr/index",
-                        "services/tandoor/index",
-                        "services/tautulli/index",
-                        "services/weblate/index",
-                        "services/youtrack/index",
-                        "services/zipline/index",
-                    ],
-                },
-                {
-                    type: "category",
-                    label: "Monitoring",
-                    items: [
-                        "services/beszel/index",
-                        "services/chronograf/index",
-                        "services/gatus/index",
-                        "services/glitchtip/index",
-                        "services/grafana/index",
-                        "services/observium/index",
-                        "services/sentry/index",
-                        "services/ubuntu-landscape/index",
-                        "services/uptime-kuma/index",
-                        "services/wazuh/index",
-                        "services/zabbix/index",
-                        "services/whats-up-docker/index",
-                    ],
-                },
-                {
-                    type: "category",
-                    label: "Platforms",
-                    items: [
-                        "services/atlassian/index",
-                        "services/budibase/index",
-                        "services/drupal/index",
-                        "services/pocketbase/index",
-                        "services/wordpress/index",
-                    ],
-                },
-                {
-                    type: "category",
-                    label: "Version Control Systems",
-                    items: [
-                        "services/gitea/index",
-                        "services/github-organization/index",
-                        "services/github-enterprise-cloud/index",
-                        "services/github-enterprise-emu/index",
-                        "services/github-enterprise-server/index",
-                        "services/gitlab/index",
-                    ],
-                },
-            ],
-        },
-    ],
-};
